+++ native/c/zut.cpp
@@ -458,6 +458,7 @@ int zut_bpxwdyn_common(const std::string &parm, unsigned int *code, std::string 
   }
 
   bparm->len = snprintf(bparm->str, sizeof(bparm->str), "%s", parm.c_str());
+
   int rc = ZUTWDYN(bparm, response);
 
   if (bparm->rtdd)
@@ -978,7 +979,7 @@ int zut_loop_dynalloc(ZDIAG &diag, const std::vector<std::string> &list)
       diag.detail_rc = ZUT_RTNCD_SERVICE_FAILURE;
       diag.service_rc = rc;
       strcpy(diag.service_name, "bpxwdyn");
-      ZDIAG_SET_MSG(&diag, "bpxwdyn failed with '%s' rc: '%d', emsg: '%s'", diag.service_name, rc, alloc.c_str());
+      ZDIAG_SET_MSG(&diag, "bpxwdyn failed with '%s' rc: '%d', emsg: '%s'", diag.service_name, rc, response.c_str());
       return RTNCD_FAILURE;
     }
   }