+++ packages/sdk/CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## Recent Changes
 
+- Added support for copying data sets and members files in the `RpcClientApi` class. [#932] (https://github.com/zowe/zowex/issues/932)
 - Fixed an issue when searching a single member did not returned properly parsed results. [#941](https://github.com/zowe/zowex/issues/941)
 - Added support for invoking the `getInfo` server command, which allows the client SDK to get version and build information from the server. [#922](https://github.com/zowe/zowex/pull/922)
 - Added warning to `AbstractConfigManager.validateDeployPath` method when server path ends in `/c/build-out`, preventing developers from accidentally overwriting a dev deployment. [#912](https://github.com/zowe/zowex/pull/912)
@@ -128,3 +129,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+